+++ src/menu.h
@@ -8,7 +8,7 @@
 extern "C" {
 #endif
 
-void menu_Looks(uint8_t *, uint8_t *, const uint8_t, const unsigned int, bool *, uint8_t *, const bool, bool *);
+void menu_Looks(uint8_t *, unsigned int *, const unsigned int, const unsigned int, bool *, uint8_t *, const bool, bool *);
 
 void menu_Info(uint8_t *, bool *, uint8_t, const unsigned int, uint8_t *, const bool, const bool);
